+++ examples/modules/partners/models/__init__.py
@@ -1,4 +1,3 @@
 from . import res_tag  # noqa: F401
 from . import res_partner  # noqa: F401
 from . import res_partner_workflow  # noqa: F401
-from . import res_country  # noqa: F401